+++ src/Main.cs
@@ -279,6 +279,7 @@ public class Initialisation : BaseUnityPlugin
                 Sunderbuss.Init();
                 Macheening.Init();
                 Stereoscope.Init();
+                Flakseed.Init();
 
                 Lazy.FinalizeGuns(); // Make sure encounter trackables are finalized so shoot styles properly display in the Ammonomicon