+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 1.22.0 (TBD)
 
-- Added Stereoscope
+- Added Stereoscope and Scalding Jelly
 - Fixed potential null deref caused by event handler for destroyed Allay companion
 
 ## 1.21.2 (2025-01-30)